render issue in product page rendering fixed
