+++ src/index.html
@@ -9,7 +9,6 @@
   </head>
   <body>
     <header class="divider" id="main-header">
-
       <!-- <div class="logo">
         <img src="/images/noun_Tent_2517.svg" alt="tent image for logo" />
         <a href="index.html"> Sleep<span class="highlight">Outside</span></a>
@@ -116,6 +115,6 @@
     </main>
     <footer id="main-footer">
       <!-- &copy;NOT a real business -->
-      </footer>
+    </footer>
   </body>
 </html>
