+++ src/js/main.js
@@ -1,8 +1,5 @@
 import { loadHeaderFooter } from "./utils.mjs";
 import productList from "./productList.mjs";
 
-
-
-productList("tents", document.querySelector(".product-list"));
 loadHeaderFooter();
-
+productList("tents", document.querySelector(".product-list"));
