+++ src/js/product.js
@@ -1,9 +1,6 @@
-import { getParam } from "./utils.mjs";
+import { getParam, loadHeaderFooter } from "./utils.mjs";
 import productDetails from "./productDetails.mjs";
-import { getCartCountFromLocalStorage } from "./utils.mjs";
 
-const backpackBadge = document.getElementById("cart-count");
-backpackBadge.innerHTML = getCartCountFromLocalStorage();
-
+loadHeaderFooter();
 const productId = getParam("product");
 productDetails(productId);
