--- src/partials/footer.html
@@ -1,1 +1,0 @@
-&copy;NOT a real business
